Changes to integrate servlets and front end

+++ unicate_server/src/main/java/csci201team25/unicate_server/Activity.java
@@ -10,11 +10,9 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 public class Activity {
 	@Id
 	@Column(name="actID")
-	private Long actID;
-	public Long getActID() {
-		return actID;
-	}
-	public void setActID(Long actID) {
+	private int actID;
+	public int getActID() { return actID; }
+	public void setActID(int actID) {
 		this.actID = actID;
 	}
